# ENH: Convert fp32 sin/cos from C universal intrinsics to C++ using Highway