--- a/numpy/_core/meson.build
+++ b/numpy/_core/meson.build
@@ -947,9 +947,7 @@ foreach gen_mtargets : [
     'src/umath/loops_trigonometric.dispatch.cpp',
     [
       AVX512_SKX, [AVX2, FMA3],
-      VSX4, VSX3, VSX2,
       NEON_VFPV4,
-      VXE2, VXE
     ]
   ],
   [