--- a/numpy/_core/src/umath/loops.h.src
+++ b/numpy/_core/src/umath/loops.h.src
@@ -10,6 +10,10 @@
     #define NPY_NO_EXPORT NPY_VISIBILITY_HIDDEN
 #endif
 
+#ifdef __cplusplus
+extern "C" {
+#endif
+
 /*
  *****************************************************************************
  **                             BOOLEAN LOOPS                               **
@@ -875,5 +879,7 @@ PyUFunc_OOO_O(char **args, npy_intp const *dimensions, npy_intp const *steps, vo
  **                              END LOOPS                                  **
  *****************************************************************************
  */
-
+#ifdef __cplusplus
+}
+#endif
 #endif