… numpy/_core/src/umath/loops_trigonometric.dispatch.c.src