This patch is to experiment with highway and see how we can leverage its intrinsics using static dispatch. I would think these are the minimum requirements: 

- [ ]  passes CI on all the relevant platforms:   AVX512_SKX, [AVX2, FMA3], VSX4, VSX3, VSX2, NEON_VFPV4, VXE2, VX. All tests pass on my local AVX512 and AVX2 machines. 
- [ ] No performance regressions. 

On x86, both AVX-512 and AVX2 seem to have performance regressions. Yet to figure out where they are coming from. 

### AVX-512 benchmarks

These are about 1.5x slower even when built with `-march=skylake-avx512`. If we use just `-mavx512f -mavx512bw`, etc, then its about 4x slower. 

```
| Change   | Before [5867ee6b] <main>   | After [d5596c17] <sincos-hwy>   |   Ratio | Benchmark (Parameter)                                                   |
|----------|----------------------------|---------------------------------|---------|-------------------------------------------------------------------------|
| +        | 7.48±0μs                   | 11.5±0.01μs                     |    1.53 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 1, 1, 'f')        |
| +        | 7.70±0.06μs                | 11.5±0.02μs                     |    1.49 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 1, 1, 'f')        |
| +        | 11.8±0.1μs                 | 16.3±0.04μs                     |    1.39 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 1, 2, 'f')        |
| +        | 11.9±0.1μs                 | 16.3±0.09μs                     |    1.38 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 1, 2, 'f')        |
| +        | 23.2±0.06μs                | 28.0±0.01μs                     |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 4, 2, 'f')        |
| +        | 23.3±0.2μs                 | 28.1±0.01μs                     |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 4, 2, 'f')        |
| +        | 20.4±0.02μs                | 24.4±0.09μs                     |    1.2  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 4, 1, 'f')        |
| +        | 20.3±0.01μs                | 24.5±0.02μs                     |    1.2  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 4, 1, 'f')        |
| -        | 66.1±0.02μs                | 62.2±0.07μs                     |    0.94 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'cos'>, 4, 2, 'f') |
| -        | 67.2±0.01μs                | 62.2±0.03μs                     |    0.93 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'sin'>, 4, 2, 'f') |
| -        | 64.6±0.03μs                | 59.1±0.01μs                     |    0.91 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'cos'>, 4, 1, 'f') |
| -        | 65.2±0.02μs                | 59.1±0.04μs                     |    0.91 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'sin'>, 4, 1, 'f') |
| -        | 60.6±0.04μs                | 54.4±0.02μs                     |    0.9  | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'cos'>, 1, 2, 'f') |
| -        | 59.0±0.05μs                | 51.6±0.04μs                     |    0.88 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'cos'>, 1, 1, 'f') |
| -        | 61.5±0.03μs                | 54.1±0.06μs                     |    0.88 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'sin'>, 1, 2, 'f') |
| -        | 59.9±0.01μs                | 51.8±0.05μs                     |    0.86 | bench_ufunc_strides.UnaryFPSpecial.time_unary(<ufunc 'sin'>, 1, 1, 'f') |
```

### AVX2 benchmarks

These are about 1.34x slower when built using `-march=skylake`. If we use `-maxv2` or even `-march=haswell`, then these seem to be 4x slower. 

```
| Change   | Before [5867ee6b] <main>   | After [d5596c17] <sincos-hwy>   |   Ratio | Benchmark (Parameter)                                            |
|----------|----------------------------|---------------------------------|---------|------------------------------------------------------------------|
| +        | 12.7±0.1μs                 | 17.1±0.1μs                      |    1.34 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 1, 1, 'f') |
| +        | 13.1±0.01μs                | 17.2±0.2μs                      |    1.31 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 1, 1, 'f') |
| +        | 40.5±0.02μs                | 49.7±0.2μs                      |    1.23 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 4, 2, 'f') |
| +        | 36.7±0.2μs                 | 45.1±0.2μs                      |    1.23 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 4, 1, 'f') |
| +        | 37.2±0.04μs                | 45.2±0.4μs                      |    1.22 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 4, 1, 'f') |
| +        | 40.3±0.01μs                | 49.3±0.2μs                      |    1.22 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 4, 2, 'f') |
| +        | 18.5±0.1μs                 | 22.2±0.2μs                      |    1.2  | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 1, 2, 'f') |
| +        | 18.5±0.02μs                | 21.8±0.2μs                      |    1.18 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 1, 2, 'f') |
```