@jan-wassenberg 2024-02-07
> -march=skylake-avx512. If we use just -mavx512f -mavx512bw, etc, then its about 4x slower.

Right, Highway checks for multiple CPU flags before using AVX3. -march=skx is sufficient here, but for the individual -mavx512, that would require a long list.