@jan-wassenberg 2024-02-29
> Moving the `hn::StoreMaskBits` to inside the if condition helped perf by a little bit, now we are just about 1.2x slower.
> 
> ```
> | +        | 7.47±0.01μs                | 9.11±0.05μs                     |    1.22 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'sin'>, 1, 1, 'f')        |
> | +        | 7.68±0.03μs                | 9.32±0.04μs                     |    1.21 | bench_ufunc_strides.UnaryFP.time_unary(<ufunc 'cos'>, 1, 1, 'f')        |
> ```

Cool, I have added some additional suggestions :)