@Mousius 2024-03-01
@jan-wassenberg just looking at the CI failures and it seems there's some compiler incompatibility on PPC (https://github.com/numpy/numpy/actions/runs/8087779139/job/22100478961?pr=25781#step:8:504) and an abort on Z13 (https://github.com/numpy/numpy/actions/runs/8087779139/job/22100479504?pr=25781) any ideas? 

There's also a failure on `armhf` but I can take a look at that when I get a minute (https://github.com/numpy/numpy/actions/runs/8087779139/job/22100478151?pr=25781)