@jan-wassenberg 2024-03-02
hm, the Z13 error is `realloc(): invalid next size
Fatal Python error: Aborted`
Seems unrelated to SIMD; are we possibly corrupting the heap?

The latter at least I can help with. We are missing HWY_ATTR:

> Additionally, each function that calls Highway ops (such as Load) must either be prefixed with HWY_ATTR, OR reside between HWY_BEFORE_NAMESPACE() and HWY_AFTER_NAMESPACE(). Lambda functions currently require HWY_ATTR before their opening brace.