@r-devulap 2024-03-20
> The latter at least I can help with. We are missing HWY_ATTR:

Adding HWY_ATTR fixed the build errors on ppc64le. Why did it fail only for this platform though? 

Need help with debugging 3 more failures:  
 
1. The [crash ](https://github.com/numpy/numpy/actions/runs/8366054700/job/22905373846?pr=25781) on s390x platforms. Ping @seiko2plus 
2. armhf failure: looks like accuracy problems ping @Mousius 
3. @jan-wassenberg any idea why the cygwin CI fails with this error? 

```
../numpy/_core/src/highway/hwy/ops/generic_ops-inl.h: In function 'void hwy::N_AVX2::StoreInterleaved3(hwy::N_AVX2::VFromD<D>, hwy::N_AVX2::VFromD<D>, hwy::N_AVX2::VFromD<D>, D, hwy::N_AVX2::TFromD<D>*)':
../numpy/_core/src/highway/hwy/ops/generic_ops-inl.h:1320:14: error: expected unqualified-id before numeric constant
 1320 |   const auto B0 = TableLookupBytesOr0(v0, shuf_B0);
```