@r-devulap 2024-03-26
> hm, strange. Neither the x86 implementation of TableLookupBytesOr0, nor the quoted line and the one before it, have a numeric constant. Which compiler is cygwin using?

From the logs: 

```
C++ compiler for the host machine: c++ (gcc 11.4.0 "c++ (GCC) 11.4.0")
C++ linker for the host machine: c++ ld.bfd 2.42
```