@jan-wassenberg 2024-03-27
hm, [here](https://gcc.godbolt.org/z/MzaGKTsqb) we are able to compile StoreInterleaved3 using GCC 11.4. Are you able to repro the issue in godbolt?