@DWesl 2024-04-04
> 3. @jan-wassenberg any idea why the cygwin CI fails with this error?
> 
> ```
> ../numpy/_core/src/highway/hwy/ops/generic_ops-inl.h: In function 'void hwy::N_AVX2::StoreInterleaved3(hwy::N_AVX2::VFromD<D>, hwy::N_AVX2::VFromD<D>, hwy::N_AVX2::VFromD<D>, D, hwy::N_AVX2::TFromD<D>*)':
> ../numpy/_core/src/highway/hwy/ops/generic_ops-inl.h:1320:14: error: expected unqualified-id before numeric constant
>  1320 |   const auto B0 = TableLookupBytesOr0(v0, shuf_B0);
> ```

Most likely `B0` is `#define`d to a number (possibly `const` or `auto`; the other places would accept a number just fine.  Could you make that name longer?