@Mousius 2024-04-06
Looks like the new AArch64 CI failure is due to missing `per_target.cc` but we really shouldn't need it if we remove the dependency on `targets.cc` - will rebase https://github.com/google/highway/pull/1935 and we can swap to only using `abort.cc` 😸