@Mousius 2024-04-08
https://github.com/google/highway/pull/2063 should fix the `armhf` failures, at least in my qemu 😸