@Mousius 2024-04-10
@r-devulap you should be able to change this line to `abort.cc`:

https://github.com/numpy/numpy/blob/0be3c26962838fb5e5fac15c3b7600f7f69f57c2/numpy/_core/meson.build#L108

That should remove all the dynamic dispatch code and fix the compilation errors 😸