@r-devulap 2024-04-11
> @r-devulap you should be able to change this line to `abort.cc`:

Done.