@jan-wassenberg 2024-04-15
FYI you can set `HWY_BASELINE_TARGETS` to your desired target e.g., HWY_AVX3, to override Highway's detection based on compiler flags.