@Mousius 2024-04-15
> FYI you can set `HWY_BASELINE_TARGETS` to your desired target e.g., (HWY_AVX3|HWY_AVX2|HWY_SSE4|HWY_EMU128), to override Highway's detection based on compiler flags.

At the moment, the constants are generated by Meson here:
https://github.com/numpy/numpy/blob/main/numpy/_core/meson.build#L778-L799

So if it finds `ASIMD`, it will add the `NPY_ASIMD` definition. What is the best way to construct `HWY_BASELINE_TARGETS` from this?