@jan-wassenberg 2024-04-15
If we also want to assume AES, then HWY_NEON, otherwise HWY_NEON_WITHOUT_AES :)