@r-devulap 2024-04-24
The sin/cos file is now free of `#include "simd/simd.h` and uses highway exclusively. I have also updated to use highway gather and scatter.