@jan-wassenberg 2024-04-25
> can someone please remind why we need to use these?

I think the reason is that Highway defaults to using whatever the build flags say to use. It is possibly easier to just set HWY_AVX2 than to set the various individual -mavx* flags that Highway would look at.